requirements/base.txt
black==23.9.1
alembic==1.12.0
fastapi[all]==0.103.1
pylint==2.17.5
passlib[bycrypt]==1.7.4
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
python-jose[cryptography]==3.3.0
python-decouple==3.8